adds man/figures/ensr.png, README.md corrections
Prior to this change the ensr.png logo was not packaged and resulted in
a WARNING on R CMD check --as-cran.  A NOTE about a url in the README.md
was also being thrown.

This change copies the logo into the man/figures/ directory and corrects
the urls in the README.md