use static html vignettes via R.rsp
Prior to this change the vignettes were built and checked by the
standard R build and check process.  The time required to re-build the
vignettes for CRAN was reported to be 17 minutes.  This was
unacceptably too long.  The build time for the vignettes, at least as
far as CRAN is concerned, needed to be drastically reduced.

This change use R.rsp as the vignette engine and creates static html
vignettes.  On my system R CMD check --as-cran took less than one
minute.

Issue #2