quick patch for "perferable" model having nzero == 0; add NEWS.md
Prior to this change if ensr::preferable returned a model with zero
non-zero (nzero) coefficients then the dimensions of the summary
data.frame had more than one row which was a problem as the code assumed
that the preferable model was unique.

This change adds a warning message and returns on model with a large
lambda value if the nzero for prefered model is zero.

re #10